HW06 jpa book info

+++ hw06-spring-jpa-book-info-app/src/main/java/ru/otus/spring/jpa_book_info_app/service/genre/GenreService.java
@@ -8,5 +8,5 @@ public interface GenreService {
 
     ServiceResult<Void> update(Genre genre);
 
-    ServiceResult<Void> remove(long id);
+    ServiceResult<Void> remove(int id);
 }
